Add files via upload
Read more about Inverse Kinematics for a 2-link arm. 
in.mathworks.com/help/symbolic/derive-and-apply-inverse-kinematics-to-robot-arm.html

This is a pre-defined MATLAB function that calculates the joint configuration, given that, you define the (x,y) coordinate of the end effector, which can be any random point in the workspace. If you want to use the function, just enter the output and input argument set for the function (here, calcIKAngles) in your own MATLAB code and paste the function at the end of the MATLAB script.

PS: Not all points in the workspace is reachable.